requirements.txt
click==6.6
cycler==0.10.0
Django==1.9.7
Flask==0.11.1
flask-marshmallow==0.6.2
Flask-Script==2.0.5
image==1.5.3
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
marshmallow==2.7.3
matplotlib==1.5.1
numpy==1.11.0
Pillow==3.2.0
psycopg2==2.6.1
pyparsing==2.1.4
python-dateutil==2.5.3
pytz==2016.4
scipy==0.17.1
six==1.10.0
Theano==0.8.2
Werkzeug==0.11.10
